refactor: move SearchFocusProvider to keyboard package
