+++ src/main/java/edu/ntnu/idatt2003/millions/view/dashboard/transactions/TransactionsView.java
@@ -2,7 +2,7 @@ package edu.ntnu.idatt2003.millions.view.dashboard.transactions;
 
 import edu.ntnu.idatt2003.millions.service.GameService;
 import edu.ntnu.idatt2003.millions.service.TransactionStatsService;
-import edu.ntnu.idatt2003.millions.view.SearchFocusProvider;
+import edu.ntnu.idatt2003.millions.keyboard.SearchFocusProvider;
 import edu.ntnu.idatt2003.millions.util.LanguageManager;
 import edu.ntnu.idatt2003.millions.view.component.StyledText;
 import edu.ntnu.idatt2003.millions.view.dashboard.transactions.card.LoanLedgerCard;
