+++ src/main/java/edu/ntnu/idatt2003/millions/view/dashboard/watchlist/WatchlistView.java
@@ -2,7 +2,7 @@ package edu.ntnu.idatt2003.millions.view.dashboard.watchlist;
 
 import edu.ntnu.idatt2003.millions.controller.TradeController;
 import edu.ntnu.idatt2003.millions.service.GameService;
-import edu.ntnu.idatt2003.millions.view.SearchFocusProvider;
+import edu.ntnu.idatt2003.millions.keyboard.SearchFocusProvider;
 import edu.ntnu.idatt2003.millions.view.component.ExploreStocksButton;
 import javafx.scene.layout.Priority;
 import javafx.scene.layout.VBox;
