+++ src/main/java/edu/ntnu/idatt2003/millions/view/leaderboard/LeaderboardView.java
@@ -3,7 +3,7 @@ package edu.ntnu.idatt2003.millions.view.leaderboard;
 import edu.ntnu.idatt2003.millions.service.GameService;
 import edu.ntnu.idatt2003.millions.service.toast.ToastService;
 import edu.ntnu.idatt2003.millions.util.LanguageManager;
-import edu.ntnu.idatt2003.millions.view.SearchFocusProvider;
+import edu.ntnu.idatt2003.millions.keyboard.SearchFocusProvider;
 import edu.ntnu.idatt2003.millions.view.component.StyledText;
 import edu.ntnu.idatt2003.millions.view.component.WeekBar;
 import edu.ntnu.idatt2003.millions.view.leaderboard.card.LeaderboardCard;
